NOTE: Been months between making changes and checking in.
Added requestStart property to SpeechRecognition to ensure that the object is reset after a defined period of time (9 seconds as defined on line 220).
Added sampleTranscription in addition to transcriptionSignal. Not sure why or what the benefits are of sampling the transcriptionSignal.
Commented print calls.
Made note about bad thread access. Not sure if this is still necessary, but leaving it in place.